requirements.txt
asgiref==3.7.2
Django==5.0.2
django-ckeditor==6.7.1
django-ckeditor-5==0.2.11
django-extensions==3.2.3
django-js-asset==2.2.0
fontawesomefree==6.5.1
pillow==10.2.0
sqlparse==0.4.4
tzdata==2023.4
